✨ feat: 기본 정보 받아오기
렌더되기 전에 훅이 여러 번 실행되는 에러 해결을 위해 컴포넌트 분리
